Renamed CausalTestOutcome to CausalEffect
