+++ causal_testing/main.py
@@ -17,7 +17,7 @@ from causal_testing.specification.variable import Input, Output
 from causal_testing.specification.causal_specification import CausalSpecification
 from causal_testing.testing.causal_test_case import CausalTestCase
 from causal_testing.testing.base_test_case import BaseTestCase
-from causal_testing.testing.causal_test_outcome import NoEffect, SomeEffect, Positive, Negative
+from causal_testing.testing.causal_effect import NoEffect, SomeEffect, Positive, Negative
 from causal_testing.testing.causal_test_result import CausalTestResult, TestValue
 from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
 from causal_testing.estimation.logistic_regression_estimator import LogisticRegressionEstimator
